eye_mouse/requirements.txt
opencv-python
mediapipe
pyautogui
playsound
pynput
Quartz


easyocr
google-generativeai==0.1.0
transformers
pillow
requests


#  pip install -r requirements.txt



